backend/requirements.txt

fastapi
uvicorn
google-generativeai
python-dotenv
pydantic
langchain
langgraph
langchain-google-genai
